# docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
  api.Error:
    properties:
      errors:
        additionalProperties: {}
        type: object
    type: object
...
  handler.GenericDataListResponse-oscalTypes_1_1_3_Finding:
    properties:
      data:
        description: Items from the list response
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Finding'
        type: array
    type: object
...
  handler.GenericDataListResponse-oscalTypes_1_1_3_Observation:
    properties:
      data:
        description: Items from the list response
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Observation'
        type: array
    type: object
  handler.GenericDataListResponse-oscalTypes_1_1_3_Party:
    properties:
      data:
        description: Items from the list response
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Party'
        type: array
    type: object
  handler.GenericDataListResponse-oscalTypes_1_1_3_PlanOfActionAndMilestones:
    properties:
      data:
        description: Items from the list response
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.PlanOfActionAndMilestones'
        type: array
    type: object
  handler.GenericDataListResponse-oscalTypes_1_1_3_PoamItem:
    properties:
      data:
        description: Items from the list response
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.PoamItem'
        type: array
    type: object
  handler.GenericDataListResponse-oscalTypes_1_1_3_Risk:
    properties:
      data:
        description: Items from the list response
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Risk'
        type: array
    type: object
...
  handler.GenericDataResponse-oscalTypes_1_1_3_Party:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/oscalTypes_1_1_3.Party'
        description: Items from the list response
    type: object
  handler.GenericDataResponse-oscalTypes_1_1_3_PlanOfActionAndMilestones:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/oscalTypes_1_1_3.PlanOfActionAndMilestones'
        description: Items from the list response
    type: object
...
  oscalTypes_1_1_3.Address:
    properties:
      addr-lines:
        items:
          type: string
        type: array
      city:
        type: string
      country:
        type: string
      postal-code:
        type: string
      state:
        type: string
      type:
        type: string
    type: object
...
  oscalTypes_1_1_3.AssociatedRisk:
    properties:
      risk-uuid:
        type: string
    type: object
...
  oscalTypes_1_1_3.BackMatter:
    properties:
      resources:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Resource'
        type: array
    type: object
...
  oscalTypes_1_1_3.Finding:
    properties:
      description:
        type: string
      implementation-statement-uuid:
        type: string
      links:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Link'
        type: array
      origins:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Origin'
        type: array
      props:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Property'
        type: array
      related-observations:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.RelatedObservation'
        type: array
      related-risks:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.AssociatedRisk'
        type: array
      remarks:
        type: string
      target:
        $ref: '#/definitions/oscalTypes_1_1_3.FindingTarget'
      title:
        type: string
      uuid:
        type: string
    type: object
  oscalTypes_1_1_3.FindingTarget:
    properties:
      description:
        type: string
      implementation-status:
        $ref: '#/definitions/oscalTypes_1_1_3.ImplementationStatus'
      links:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Link'
        type: array
      props:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Property'
        type: array
      remarks:
        type: string
      status:
        $ref: '#/definitions/oscalTypes_1_1_3.ObjectiveStatus'
      target-id:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  oscalTypes_1_1_3.ImportSsp:
    properties:
      href:
        type: string
      remarks:
        type: string
    type: object
...
  oscalTypes_1_1_3.Link:
    properties:
      href:
        type: string
      media-type:
        type: string
      rel:
        type: string
      resource-fragment:
        type: string
      text:
        type: string
    type: object
...
  oscalTypes_1_1_3.Metadata:
    properties:
      actions:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Action'
        type: array
      document-ids:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.DocumentId'
        type: array
      last-modified:
        type: string
      links:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Link'
        type: array
      locations:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Location'
        type: array
      oscal-version:
        type: string
      parties:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Party'
        type: array
      props:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Property'
        type: array
      published:
        type: string
      remarks:
        type: string
      responsible-parties:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.ResponsibleParty'
        type: array
      revisions:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.RevisionHistoryEntry'
        type: array
      roles:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Role'
        type: array
      title:
        type: string
      version:
        type: string
    type: object
...
  oscalTypes_1_1_3.Observation:
    properties:
      collected:
        type: string
      description:
        type: string
      expires:
        type: string
      links:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Link'
        type: array
      methods:
        items:
          type: string
        type: array
      origins:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Origin'
        type: array
      props:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Property'
        type: array
      relevant-evidence:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.RelevantEvidence'
        type: array
      remarks:
        type: string
      subjects:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.SubjectReference'
        type: array
      title:
        type: string
      types:
        items:
          type: string
        type: array
      uuid:
        type: string
    type: object
...
  oscalTypes_1_1_3.Origin:
    properties:
      actors:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.OriginActor'
        type: array
      related-tasks:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.RelatedTask'
        type: array
    type: object
...
  oscalTypes_1_1_3.Party:
    properties:
      addresses:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Address'
        type: array
      email-addresses:
        items:
          type: string
        type: array
      external-ids:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.PartyExternalIdentifier'
        type: array
      links:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Link'
        type: array
      location-uuids:
        items:
          type: string
        type: array
      member-of-organizations:
        items:
          type: string
        type: array
      name:
        type: string
      props:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Property'
        type: array
      remarks:
        type: string
      short-name:
        type: string
      telephone-numbers:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.TelephoneNumber'
        type: array
      type:
        type: string
      uuid:
        type: string
    type: object
  oscalTypes_1_1_3.PartyExternalIdentifier:
    properties:
      id:
        type: string
      scheme:
        type: string
    type: object
  oscalTypes_1_1_3.PlanOfActionAndMilestones:
    properties:
      back-matter:
        $ref: '#/definitions/oscalTypes_1_1_3.BackMatter'
      findings:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Finding'
        type: array
      import-ssp:
        $ref: '#/definitions/oscalTypes_1_1_3.ImportSsp'
      local-definitions:
        $ref: '#/definitions/oscalTypes_1_1_3.PlanOfActionAndMilestonesLocalDefinitions'
      metadata:
        $ref: '#/definitions/oscalTypes_1_1_3.Metadata'
      observations:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Observation'
        type: array
      poam-items:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.PoamItem'
        type: array
      risks:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Risk'
        type: array
      system-id:
        $ref: '#/definitions/oscalTypes_1_1_3.SystemId'
      uuid:
        type: string
    type: object
  oscalTypes_1_1_3.PlanOfActionAndMilestonesLocalDefinitions:
    properties:
      assessment-assets:
        $ref: '#/definitions/oscalTypes_1_1_3.AssessmentAssets'
      components:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.SystemComponent'
        type: array
      inventory-items:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.InventoryItem'
        type: array
      remarks:
        type: string
    type: object
  oscalTypes_1_1_3.PoamItem:
    properties:
      description:
        type: string
      links:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Link'
        type: array
      origins:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.PoamItemOrigin'
        type: array
      props:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Property'
        type: array
      related-findings:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.RelatedFinding'
        type: array
      related-observations:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.RelatedObservation'
        type: array
      related-risks:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.AssociatedRisk'
        type: array
      remarks:
        type: string
      title:
        type: string
      uuid:
        type: string
    type: object
...
  oscalTypes_1_1_3.Property:
    properties:
      class:
        type: string
      group:
        type: string
      name:
        type: string
      ns:
        type: string
      remarks:
        type: string
      uuid:
        type: string
      value:
        type: string
    type: object
...
  oscalTypes_1_1_3.RelatedObservation:
    properties:
      observation-uuid:
        type: string
    type: object
...
  oscalTypes_1_1_3.RelevantEvidence:
    properties:
      description:
        type: string
      href:
        type: string
      links:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Link'
        type: array
      props:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Property'
        type: array
      remarks:
        type: string
    type: object
...
  oscalTypes_1_1_3.Risk:
    properties:
      characterizations:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Characterization'
        type: array
      deadline:
        type: string
      description:
        type: string
      links:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Link'
        type: array
      mitigating-factors:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.MitigatingFactor'
        type: array
      origins:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Origin'
        type: array
      props:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Property'
        type: array
      related-observations:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.RelatedObservation'
        type: array
      remediations:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Response'
        type: array
      risk-log:
        $ref: '#/definitions/oscalTypes_1_1_3.RiskLog'
      statement:
        type: string
      status:
        type: string
      threat-ids:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.ThreatId'
        type: array
      title:
        type: string
      uuid:
        type: string
    type: object
...
  oscalTypes_1_1_3.SubjectReference:
    properties:
      links:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Link'
        type: array
      props:
        items:
          $ref: '#/definitions/oscalTypes_1_1_3.Property'
        type: array
      remarks:
        type: string
      subject-uuid:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  oscalTypes_1_1_3.SystemId:
    properties:
      id:
        type: string
      identifier-type:
        type: string
    type: object
...
  oscalTypes_1_1_3.TelephoneNumber:
    properties:
      number:
        type: string
      type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API for the Continuous Compliance Framework.
  title: Continuous Compliance Framework API
  version: "1"
paths:
...
  /oscal/parties:
    get:
      description: Retrieves all parties.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataListResponse-oscalTypes_1_1_3_Party'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Password: []
      summary: List parties
      tags:
      - Oscal
  /oscal/parties/{id}:
    get:
      description: Retrieves a single Party by its unique ID.
      parameters:
      - description: Party ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataResponse-oscalTypes_1_1_3_Party'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Password: []
      summary: Get a Party
      tags:
      - Oscal
  /oscal/plan-of-action-and-milestones:
    get:
      description: Retrieves all Plan of Action and Milestones.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataListResponse-oscalTypes_1_1_3_PlanOfActionAndMilestones'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: List POA&Ms
      tags:
      - OScal
  /oscal/plan-of-action-and-milestones/{id}:
    get:
      description: Retrieves a single Plan of Action and Milestones by its unique
        ID.
      parameters:
      - description: POA&M ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataResponse-oscalTypes_1_1_3_PlanOfActionAndMilestones'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a POA&M
      tags:
      - OScal
  /oscal/plan-of-action-and-milestones/{id}/findings:
    get:
      description: Retrieves all findings for a given POA&M.
      parameters:
      - description: POA&M ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataListResponse-oscalTypes_1_1_3_Finding'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get findings for a POA&M
      tags:
      - OScal
  /oscal/plan-of-action-and-milestones/{id}/full:
    get:
      description: Retrieves a complete POA&M by its ID, including all metadata and
        related objects.
      parameters:
      - description: POA&M ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataResponse-oscalTypes_1_1_3_PlanOfActionAndMilestones'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get a complete POA&M
      tags:
      - OScal
  /oscal/plan-of-action-and-milestones/{id}/observations:
    get:
      description: Retrieves all observations for a given POA&M.
      parameters:
      - description: POA&M ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataListResponse-oscalTypes_1_1_3_Observation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get observations for a POA&M
      tags:
      - OScal
  /oscal/plan-of-action-and-milestones/{id}/poam-items:
    get:
      description: Retrieves all POA&M items for a given POA&M.
      parameters:
      - description: POA&M ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataListResponse-oscalTypes_1_1_3_PoamItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get POA&M items
      tags:
      - OScal
  /oscal/plan-of-action-and-milestones/{id}/risks:
    get:
      description: Retrieves all risks for a given POA&M.
      parameters:
      - description: POA&M ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GenericDataListResponse-oscalTypes_1_1_3_Risk'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Error'
      summary: Get risks for a POA&M
      tags:
      - OScal
...
produces:
- application/json
securityDefinitions:
  OAuth2Password:
    flow: password
    tokenUrl: /api/auth/token
    type: oauth2
swagger: "2.0"